requirements.txt
tensorflow-cpu~=2.11.0
streamlit~=1.19.0
numpy~=1.24.2
Pillow~=9.4.0